@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp2.2</TargetFramework>
+        <TargetFramework>netcoreapp5.0</TargetFramework>
         <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
     </PropertyGroup>
 
@@ -11,10 +11,12 @@
         <PackageReference Include="Hangfire.Core" Version="1.7.9" />
         <PackageReference Include="Hangfire.SqlServer" Version="1.7.9" />
         <PackageReference Include="Microsoft.AspNetCore.App" />
+        <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="5.0.3" />
+        <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="5.0.3" />
         <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
-        <PackageReference Include="Microsoft.EntityFrameworkCore" Version="2.2.2" />
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="2.2.2" />
-        <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="2.2.2" />
+        <PackageReference Include="Microsoft.EntityFrameworkCore" Version="5.0.3" />
+        <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="5.0.3" />
+        <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="5.0.3" />
     </ItemGroup>
     
     <ItemGroup>
